convoy driver: Add cleanup option for delete volume
If we want to use NFS, we probably would end up with one volume directory
reference by different volumes across different hosts, if user want to move the
volume. We need to have a way to only delete the volume reference on local
host, but not the data directory.

This commit introduce "--cleanup" for delete volume, which would pass to VFS
driver, and data directory of VFS's volume would only be deleted if "--cleanup"
was passed.