Add driver option for supporting xfs #84
This change adds a new driver option dm.fs for devmapper to support
formatting devices with filesystem types specified by user. So far,
supported filesystems are ext4 and xfs. It defaults to ext4 if the
option is specified.

Integration test passed as well as manual test with the new dm.fs option.

Signed-off-by: Jin Xu <jinuxstyle@hotmail.com>