Add option to ignore existing config file
When the config file exists, convoy daemon always loads it and ignore
user's command line options. The only way out is deleting the config
file manually. This is not user friendly if user would like to start
the daemon with different options from the originally saved. So a
new option --ignore-config-file is added to avoid using the existing
config file.

Signed-off-by: Jin Xu <jinuxstyle@hotmail.com>