Add option to ignore existing config file #108
Convoy daemon saves command line options to configure files for the
first run. Later on when user starts the daemon again with different
command line options, it will always load the configure files and
use the previously saved command lines options instead of using the
new options. The only way out is deleting the config file manually or
specifing a different directory for --root option.

This patch adds a new option --ignore-config-file to fix the problem.
When this option is given, convoy daemon will ignore the general
configure file and ignore the driver specific configure file only if
there is no existing volume under the driver.

Signed-off-by: Jin Xu <jinuxstyle@hotmail.com>