Add configurable timeout option for executing commands #73,#99
Some operations are too time-consuming to finish in the hard coded
one minute. For example, it causes timeout and error when creating
a new volume (#73) or snapshoting large volumes (#99). This patch
makes the timeout value confirable by adding a new option cmd-timeout.
It defaults to one minute if user passes an invalid value or doesn't
set the value. The value should be a duration string. For example,
setting to ten minutes by "10m" and one hour by "1h". For more valid
time units, refer to the manual of time.ParseDuration.

Signed-off-by: Jin Xu <jinuxstyle@hotmail.com>