CLOUD-298 Convoy should infer volume size from name
This commit enables convoy to read the volume name and create
the volume as the expected default behavior, but also scan the
name for `@XG` appended to the name where XG is size in gigabytes.

The volume is created of this specific size, while this data is
consumed by convoy so that we leave no traces of it in the backend.
This makes this change backwards compatible as well as easy to deprecate
in the future.